Conflicts:
	examples/mdp/asynchronous_leader/leader3.tra
	examples/mdp/asynchronous_leader/leader4.tra
	examples/mdp/asynchronous_leader/leader5.tra
	examples/mdp/asynchronous_leader/leader6.tra
	examples/mdp/asynchronous_leader/leader7.tra
	examples/mdp/two_dice/two_dice.lab
	examples/mdp/two_dice/two_dice.tra
	src/formula/NoBoundOperator.h
	src/formula/PathBoundOperator.h
	src/formula/ProbabilisticNoBoundOperator.h
	src/formula/RewardNoBoundOperator.h
	src/modelchecker/AbstractModelChecker.h
	src/modelchecker/DtmcPrctlModelChecker.h
	src/modelchecker/EigenDtmcPrctlModelChecker.h
	src/modelchecker/GmmxxDtmcPrctlModelChecker.h
	src/modelchecker/GmmxxMdpPrctlModelChecker.h
	src/modelchecker/MdpPrctlModelChecker.h
	src/models/GraphTransitions.h
	src/parser/AutoParser.h
	src/parser/DeterministicSparseTransitionParser.cpp
	src/parser/NondeterministicModelParser.cpp
	src/parser/NondeterministicSparseTransitionParser.cpp
	src/parser/NondeterministicSparseTransitionParser.h
	src/storage/BitVector.h
	src/storage/SparseMatrix.h
	src/storm.cpp
	src/utility/GraphAnalyzer.h
	src/utility/Vector.h

+++ src/adapters/EigenAdapter.h
@@ -32,9 +32,9 @@ public:
 		LOG4CPLUS_DEBUG(logger, "Converting matrix with " << realNonZeros << " non-zeros to Eigen format.");
 
 		// Prepare the resulting matrix.
-		Eigen::SparseMatrix<T, Eigen::RowMajor, int_fast32_t>* result = new Eigen::SparseMatrix<T, Eigen::RowMajor, int_fast32_t>(matrix.rowCount, matrix.colCount);
+		Eigen::SparseMatrix<T, Eigen::RowMajor, int_fast32_t>* result = new Eigen::SparseMatrix<T, Eigen::RowMajor, int_fast32_t>(static_cast<int>(matrix.rowCount), static_cast<int>(matrix.colCount));
 
-		result->resizeNonZeros(realNonZeros);
+		result->resizeNonZeros(static_cast<int>(realNonZeros));
 		//result->reserve(realNonZeros);
 
 		// Copy Row Indications
@@ -54,4 +54,4 @@ public:
 
 } //namespace storm
 
-#endif /* STORM_ADAPTERS_GMMXXADAPTER_H_ */
+#endif /* STORM_ADAPTERS_EIGENADAPTER_H_ */
